mcp_tut/pyproject.toml
[project]
name = "mcp-tut"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "cachetools>=6.1.0",
    "clerk-sdk>=0.3.0",
    "dotenv>=0.9.9",
    "fastmcp>=2.11.3",
    "httpx>=0.28.1",
    "python-jose>=3.5.0",
    "sqlalchemy>=2.0.43",
    "stytch>=13.19.0",
]
